requirements.txt
aiogram==2.19
tortoise-orm==0.18.1
asyncpg==0.25.0